requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
bcrypt<4.0
